+++ leif/leifsettings.h
@@ -23,14 +23,14 @@ public:
     QLocale::Country country() const;
     QString regionId() const;
 
-    void saveLifetimeCarbon(int lifeTime);
-    int lifeTimeCarbon() const;
+    void saveLifetimeCarbon(float lifeTime);
+    float lifeTimeCarbon() const;
     void clearLifeTimeCarbon();
 
 signals:
     void countryChanged(const QLocale::Country country);
     void regionIdChanged(const QString &regionId);
-    void lifeTimeCarbonChanged(int lifeTime);
+    void lifeTimeCarbonChanged(float lifeTime);
 
 private:
     Q_DISABLE_COPY_MOVE(LeifSettings);
